Extract the script from MakersIndexView.vue into a composable
